Handle directives whose names overlap.
A user may define a directive @@date and another directive @@datenow. When
@@date is searched, @@datenow should be skipped. A check was introduced into
pandoc-process-directives to enable this.

As a result, directive names should only consist of letters and numbers. No such
check is performed when defining directives, however. (And in fact, most of the
time things will run without problems if the user should define such a
directive.) Perhaps an explicit check should be introduced.