Modify behaviour of pandoc-run-markdown2pdf.
When called without prefix argument, pandoc-run-markdown2pdf now uses the
settings of the current buffer if its output format is "latex", rather than read
the settings from the LaTeX settings file. If the buffer's output format is not
"latex", or if pandoc-run-markdown2pdf is called with a prefix argument, the
LaTeX settings file is used, or if none is found, all options except for the
input and output format are unset.