Process all @@-directives in one go.
Instead of searching for @@-directives in the order in which they appear
in `pandoc-directives`, use a regexp to search for all of them and
process them all in one go.

This makes it possible to have directives produce other directives
without taking into account the order in which they are listed in
`pandoc-directives`.