Update for Pandoc version 2.12
- Add `org` as a possible format for the extension `task_list`.
- Change `resource-path` from a string option to a list option.

The latter change will cause an error for users that use this option, but the
code isn't very amenable to adding a check for the type of the `resource-path`
option's value. There is no clear place in the code where it could be added. The
logical place would be `pandoc--read-settings-from-file`, but it reads the
settings with `read` in one go.