Remove font-lock faces for citations save for the key.
The existing regexps are too greedy, see Github issue #105. A closer look
revealed that in citations with multiple keys, they actually match the wrong
thing: only one key is really matched as a key, the others are matched as extra
material (prefix, locator or suffix in CSL parlance).

Fixing this is difficult,[1] so it seems best to just font-lock the key itself
for now so that at least issue #105 is fixed. We can always revisit the problem
later.

[1] <https://lists.gnu.org/archive/html/help-gnu-emacs/2021-03/msg00104.html>