`pandoc--compose-output-file-name`: Do not expand file path.
We now use this function in `pandoc--pp-option` for the `output` option,
and we want to show the path that is actually stored in the settings, which
is normally a non-expanded path.

Also use `font-lock-comment-face` for those parts of the file path that are
inferred based on the input file path.