`pandoc--compose-output-file-naem`: Add argument `no-expand`.
This function is supposed to return nil if it cannot create at output file
name (mainly when the `output-file` option is unset, which means to pass
output to stdout). We cannot pass nil to `expand-file-name`, however, so
it is easier to let `pandoc--compose-output-file-name` to do the expansion.
We prefer to display the non-expanded version in the transient,
however, because it better reflects what the actual setting is, so we make expansion optional.

t# with '#' will be ignored, and an empty message aborts the commit.