Read file paths relative to `resource-path`, if possible.
When reading a file path, if the file is in (a subdirectory of) one of the
directories in the option `resource-path`, return a relative path starting
from this directory. Otherwise return a file path relative to
`default-directory`.

Pandoc searches for files in `resourse-path`, or starting from the working
directory if `resource-path` is empty. This change should reflect that behaviour.