Correctly check if `pandoc--call-pandoc` should send the buffer to pandoc.
If `input-files` is set, `pandoc--call-pandoc` should not pass the buffer
text to pandoc, rather the input file or files are in the defaults file. If
the user uses a different output format than the one set for the current
buffer, we need to check the settings of the corresponding settings file,
not the settings for the current buffer.